scripts/requirements.txt
paho-mqtt==1.6.1
pandas==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
